+++ modules/qml_engine/builtin/cpp/qt_constantBuiltin.cpp
@@ -1,0 +1,60 @@
+//=============================================================================
+// Copyright (c) 2016-present Allan CORNET (Nelson)
+//=============================================================================
+// This file is part of the Nelson.
+//=============================================================================
+// LICENCE_BLOCK_BEGIN
+// This program is free software; you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public
+// License as published by the Free Software Foundation; either
+// version 2.1 of the License, or (at your option) any later version.
+//
+// Alternatively, you can redistribute it and/or
+// modify it under the terms of the GNU General Public License as
+// published by the Free Software Foundation; either version 2 of
+// the License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public
+// License along with this program. If not, see <http://www.gnu.org/licenses/>.
+// LICENCE_BLOCK_END
+//=============================================================================
+#include "qt_constantBuiltin.hpp"
+#include "QtConstants.hpp"
+#include "Error.hpp"
+#include "ToCellString.hpp"
+//=============================================================================
+using namespace Nelson;
+//=============================================================================
+ArrayOfVector
+Nelson::QmlEngineGateway::qt_constantBuiltin(int nLhs, const ArrayOfVector& argIn)
+{
+    if (nLhs > 1) {
+        Error(ERROR_WRONG_NUMBERS_OUTPUT_ARGS);
+    }
+    ArrayOfVector retval;
+    switch (argIn.size()) {
+    case 0: {
+        retval.push_back(ToCellStringAsColumn(QtConstants()));
+    } break;
+    case 1: {
+        std::wstring constantName = argIn[0].getContentAsWideString();
+        bool found;
+        ArrayOf res = QtConstant(constantName, found);
+        if (!found) {
+            Error(_W("Name not found."));
+        } else {
+            retval.push_back(res);
+        }
+    } break;
+    default: {
+        Error(ERROR_WRONG_NUMBERS_INPUT_ARGS);
+    } break;
+    }
+    return retval;
+}
+//=============================================================================